+++ src/main/scala/view/View.scala
@@ -1,5 +1,5 @@
 package view
 
 trait View[T] {
-  def render(model: T): String
+  def update(model: T): String
 }
